backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
requests
pyjwt
numpy
pandas
pydantic
sqlalchemy
scikit-learn
joblib
passlib[bcrypt]
